Replaced React image with Chackra image to fix server error
Checking for a nonexistent image externally on the server side makes no sense, so we use the Chackra Image which pulls on client-side
thereby preventing those image not found server errors. Also you can specify an alternate image, if the image is not found anymore, which is
set to the default battle.net image. We should avoid using the react image, since I think it is meant for self-hosted images.

Signed-off-by: Benedek Major <benedek@major.onl>